ruleset: Rename set_no_new_privs() and extract NNP helper
The Rust builder pattern is meant to read as a declarative list of
configurations rather than a sequence of imperative commands, see
https://rust-lang.github.io/api-guidelines/type-safety.html?#builders-enable-construction-of-complex-values-c-builder
https://rust-lang.github.io/rfcs/0344-conventions-galore.html#gettersetter-apis

Dropping set_ from boolean field setters reduces boilerplate and makes the
chain read uniformly (with upcoming log_* setters):

    Ruleset::default()
        .handle_access(...)?
        .create()?
        .log_same_exec(false)?
        .log_new_exec(true)?
        .log_subdomains(false)?
        .no_new_privs(false)
        .restrict_self()?;

Note that this method should only be rarely called because of the
default value.

Verb-led method names are reserved for distinct actions beyond field
assignment (handle_access, add_rule, restrict_self).

Keep set_no_new_privs as a #[deprecated] thin wrapper calling
no_new_privs() to preserve backward compatibility.

set_compatibility() is out of scope for now, see
https://github.com/landlock-lsm/rust-landlock/pull/58

Also extract the prctl(PR_SET_NO_NEW_PRIVS) + CompatLevel error
dispatch from RulesetCreated::restrict_self() into a free function
try_set_no_new_privs(&mut Compatibility) -> Result<bool, RestrictSelfError>
in a new prctl module that also holds the pre-existing
prctl_set_no_new_privs() and support_no_new_privs() wrappers.  The
set_ prefix is retained because the helper performs a fixed action
(always sets PR_SET_NO_NEW_PRIVS to 1) rather than configuring a
value; the bare-name rule applies to builder setters that take a
value to configure.  This groups all prctl(2) plumbing into one file
and prepares the helper for sharing with a future caller (the
RestrictSelf builder added in a following commit).
